cut1 file used is by default the local one
git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/ILCDIRAC/trunk/ILCDIRAC@32245 4525493e-7705-40b1-a816-d608a930855b

+++ Workflow/Modules/WhizardAnalysis.py
@@ -287,7 +287,8 @@ class WhizardAnalysis(ModuleBase):
       for gridfile in list_of_gridfiles:
         script.write('ln -s %s/%s\n'%(path_to_gridfiles,gridfile))
     script.write('ln -s %s/whizard.prc\n'%mySoftDir)
-    script.write('ln -s %s/whizard.cut1\n'%mySoftDir)
+    if not os.path.exists("whizard.cut1"):
+      script.write('ln -s %s/whizard.cut1\n'%mySoftDir)
     script.write('echo =============================\n')
     script.write('echo Printing content of whizard.prc \n')
     script.write('cat whizard.prc\n')
